+++ stencil.config.ts
@@ -18,6 +18,8 @@ import { getExcludedComponentTags } from './src/global/scripts/exclude-react-com
  */
 const excludeComponents = getExcludedComponentTags('./src/components/functional');
 
+const isDev = process.argv.includes('--dev');
+
 export const config: Config = {
   namespace: 'sdk-dapp-ui',
   globalScript: './src/global/scripts/fonts-loader.ts',
@@ -27,6 +29,7 @@ export const config: Config = {
       tailwindCssPath: './src/global/tailwind.css',
     }),
   ],
+  sourceMap: isDev,
   outputTargets: [
     reactOutputTarget({
       outDir: './dist/react',
@@ -45,11 +48,6 @@ export const config: Config = {
       copy: [{ src: 'assets', dest: 'assets' }],
       esmLoaderPath: './loader',
     },
-    // this is only for testing purposes
-    // {
-    //   type: 'www',
-    //   serviceWorker: null,
-    // },
   ],
   rollupPlugins: {
     before: [nodePolyfills()],